chore: Create `useGeolocation` hook
